Split go/scheduler into scheduler, query
The scheduler.go file was always schizophrenic as it exposed a
"scheduler" interface which both scheduled (implies I/O, or at the very
least queueing) and planned queries. This feature has been split into
its scheduling and planning parts, and the query/plan stuff has seen
some minor renaming to emphasise this.
